backend/function-app/requirements.txt
./vendor/lna_db-0.1.0-py3-none-any.whl
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
beanie==1.29.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
coverage==7.6.1 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
lazy-model==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
motor==3.6.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.3 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.9.2 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==0.24.0 ; python_version >= "3.11" and python_version < "4.0"
pytest-cov==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.3.5 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
